@@ -2,11 +2,11 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId 'org.xiaoxingqi.gmdoc'
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -48,15 +48,15 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     //    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     api 'org.jetbrains.anko:anko-sdk15:0.8.2'
     // sdk19, sdk21, sdk23 are also available
     api 'org.jetbrains.anko:anko-support-v4:0.8.2'
     // In case you need support-v4 bindings
     api 'org.jetbrains.anko:anko-appcompat-v7:0.8.2'
     // For appcompat-v7 bindings
-    api 'com.android.support:design:28.0.0'
+    api 'com.google.android.material:material:1.0.0'
     api 'com.squareup.okhttp3:logging-interceptor:3.1.2'
     api 'com.squareup.okhttp3:okhttp:3.4.1'
     api 'com.jakewharton:disklrucache:2.0.2'
@@ -71,7 +71,7 @@ dependencies {
     api 'org.greenrobot:eventbus:3.0.0'
     api 'com.github.bumptech.glide:glide:3.6.1'
     api 'com.github.bumptech.glide:okhttp3-integration:1.4.0@aar'
-    api 'com.android.support:cardview-v7:28.0.0'
+    api 'androidx.cardview:cardview:1.0.0'
     implementation 'com.nineoldandroids:library:2.4.0'
     implementation files('libs/universal-image-loader-1.9.4-with-sources.jar')
     api 'me.everything:overscroll-decor-android:1.0.4'
