@@ -1,6 +1,6 @@
 package org.xiaoxingqi.gmdoc.core.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.SparseArray;
 import android.view.View;
 import android.widget.Button;
