@@ -1,7 +1,7 @@
 package org.xiaoxingqi.gmdoc.modul.game
 
 import android.content.Intent
-import android.support.v7.widget.LinearLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.text.TextUtils
 import android.view.View
 import com.bumptech.glide.Glide
