@@ -2,7 +2,7 @@ package org.xiaoxingqi.gmdoc.modul.game
 
 import android.app.Activity
 import android.content.Intent
-import android.support.v7.widget.GridLayoutManager
+import androidx.recyclerview.widget.GridLayoutManager
 import android.text.TextUtils
 import android.view.View
 import android.widget.ImageView
