@@ -1,7 +1,7 @@
 package org.xiaoxingqi.gmdoc.modul.home
 
 import android.graphics.Color
-import android.support.v7.widget.LinearLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.ImageView
