@@ -5,14 +5,17 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.support.v4.app.*
-import android.support.v4.view.ViewPager
+import androidx.core.app.*
+import androidx.viewpager.widget.ViewPager
 import android.text.TextUtils
 import android.transition.ArcMotion
 import android.util.TypedValue
 import android.view.View
 import android.view.animation.AnimationUtils
 import android.widget.ImageView
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
 import com.bumptech.glide.Glide
 import kotlinx.android.synthetic.main.activity_user_info.*
 import org.xiaoxingqi.gmdoc.R
