@@ -1,6 +1,6 @@
 package org.xiaoxingqi.gmdoc.modul.home
 
-import android.support.v7.widget.GridLayoutManager
+import androidx.recyclerview.widget.GridLayoutManager
 import android.view.LayoutInflater
 import android.view.View.GONE
 import android.view.View.VISIBLE
