@@ -1,8 +1,8 @@
 package org.xiaoxingqi.gmdoc.modul.message
 
 import android.content.Intent
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import com.bumptech.glide.Glide
@@ -69,7 +69,7 @@ class MsgFragment : BaseFrag<MessagePresenter>() {
             }
         }
         recyclerView.adapter = adapter
-        persent?.getMsgList()
+        presenter?.getMsgList()
     }
 
     override fun bindEvent() {
