@@ -3,7 +3,7 @@ package org.xiaoxingqi.gmdoc.wegidt;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Paint;
-import android.support.v7.widget.AppCompatTextView;
+import androidx.appcompat.widget.AppCompatTextView;
 import android.util.AttributeSet;
 
 import org.xiaoxingqi.gmdoc.R;
