@@ -1,7 +1,7 @@
 package org.xiaoxingqi.gmdoc.wegidt;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageView;
+import androidx.appcompat.widget.AppCompatImageView;
 import android.util.AttributeSet;
 
 /**
