@@ -1,19 +1,19 @@
 package org.xiaoxingqi.gmdoc.wegidt.gameDetails;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageView;
+import androidx.appcompat.widget.AppCompatImageView;
 import android.util.AttributeSet;
 
-public class SqureImageView extends AppCompatImageView {
-    public SqureImageView(Context context) {
+public class SquareImageView extends AppCompatImageView {
+    public SquareImageView(Context context) {
         super(context);
     }
 
-    public SqureImageView(Context context, AttributeSet attrs) {
+    public SquareImageView(Context context, AttributeSet attrs) {
         super(context, attrs);
     }
 
-    public SqureImageView(Context context, AttributeSet attrs, int defStyleAttr) {
+    public SquareImageView(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
     }
 
