@@ -2,8 +2,8 @@ package org.xiaoxingqi.gmdoc.wegidt.gamefraglist
 
 import android.content.Context
 import android.graphics.Color
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.util.AttributeSet
 import android.util.Log
 import android.view.View
