@@ -10,8 +10,8 @@ import android.graphics.RectF;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.support.v4.view.MotionEventCompat;
-import android.support.v7.widget.AppCompatTextView;
+import androidx.core.view.MotionEventCompat;
+import androidx.appcompat.widget.AppCompatTextView;
 import android.text.Html;
 import android.text.Layout;
 import android.text.Spannable;
@@ -23,7 +23,6 @@ import android.text.style.BackgroundColorSpan;
 import android.text.style.ForegroundColorSpan;
 import android.text.style.ImageSpan;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewConfiguration;
