@@ -3,7 +3,7 @@ package org.xiaoxingqi.gmdoc.wegidt.textView;
 import android.content.Context;
 import android.content.res.AssetManager;
 import android.graphics.drawable.BitmapDrawable;
-import android.support.v7.widget.AppCompatTextView;
+import androidx.appcompat.widget.AppCompatTextView;
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
 import android.util.AttributeSet;
